split equation solving into a separate file
